# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Amiga Specialty HR Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /storage/uploads/request-url:
    post:
      operationId: requestUploadUrl
      tags: [storage]
      summary: Request a presigned upload URL
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RequestUploadUrlBody" }
      responses:
        "200":
          description: Upload URL
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RequestUploadUrlResponse" }
  /jobs:
    get:
      operationId: listJobs
      tags: [jobs]
      parameters:
        - in: query
          name: status
          schema: { type: string, enum: [open, closed, draft] }
      responses:
        "200":
          description: Jobs
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Job" }
    post:
      operationId: createJob
      tags: [jobs]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateJobBody" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Job" }
...
  /public/jobs:
    get:
      operationId: listPublicJobs
      tags: [public]
      responses:
        "200":
          description: Open jobs
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Job" }
        "503":
          description: Open roles are temporarily unavailable
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PublicJobsUnavailableResponse" }
  /public/applications:
    post:
      operationId: submitApplication
      tags: [public]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PublicApplicationBody" }
      responses:
        "200":
          description: Existing application returned for an idempotent retry
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PublicApplicationResponse" }
        "201":
          description: Submitted
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PublicApplicationResponse" }
  /public/storage/uploads/request-url:
    post:
      operationId: requestPublicUploadUrl
      tags: [public]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RequestUploadUrlBody" }
      responses:
        "200":
          description: Upload URL
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RequestUploadUrlResponse" }
...
components:
  schemas:
...
    RequestUploadUrlBody:
      type: object
      properties:
        name: { type: string }
        size: { type: integer }
        contentType: { type: string }
      required: [name, size, contentType]
    RequestUploadUrlResponse:
      type: object
      properties:
        uploadURL: { type: string }
        objectPath: { type: string }
      required: [uploadURL, objectPath]
    Job:
      type: object
      properties:
        id: { type: integer }
        title: { type: string }
        department: { type: string }
        location: { type: string }
        employmentType:
          type: string
          enum: [full_time, part_time, contractor]
        salaryMin: { type: ["number", "null"] }
        salaryMax: { type: ["number", "null"] }
        currency: { type: string }
        description: { type: ["string", "null"] }
        status:
          type: string
          enum: [open, closed, draft]
        closingDate: { type: ["string", "null"] }
        hiringManagerId: { type: ["integer", "null"] }
        candidateCount: { type: integer }
        hiredCount: { type: integer }
        createdAt: { type: string }
        updatedAt: { type: string }
      required:
        [id, title, department, location, employmentType, currency, status, candidateCount, hiredCount, createdAt, updatedAt]
    CreateJobBody:
      type: object
      properties:
        title: { type: string }
        department: { type: string }
        location: { type: ["string", "null"] }
        employmentType:
          type: string
          enum: [full_time, part_time, contractor]
        salaryMin: { type: ["number", "null"] }
        salaryMax: { type: ["number", "null"] }
        currency: { type: ["string", "null"] }
        description: { type: ["string", "null"] }
        status:
          type: string
          enum: [open, closed, draft]
        closingDate: { type: ["string", "null"] }
        hiringManagerId: { type: ["integer", "null"] }
      required: [title, department, employmentType]
...
    CreateCandidateDocumentBody:
      type: object
      properties:
        name: { type: string }
        category:
          type: string
          enum: [cv, cover_letter, passport, right_to_work, qualification, reference, other]
        objectPath: { type: string }
        fileSize: { type: ["integer", "null"] }
        mimeType: { type: ["string", "null"] }
      required: [name, category, objectPath]
...
    PublicApplicationBody:
      type: object
      properties:
        firstName: { type: string }
        lastName: { type: string }
        email: { type: string }
        phone: { type: ["string", "null"] }
        jobId: { type: ["integer", "null"] }
        linkedinUrl: { type: ["string", "null"] }
        currentCompany: { type: ["string", "null"] }
        currentRole: { type: ["string", "null"] }
        yearsExperience: { type: ["integer", "null"] }
        currentSalary: { type: ["number", "null"] }
        expectedSalary: { type: ["number", "null"] }
        noticePeriod: { type: ["string", "null"] }
        lastBonusAmount: { type: ["number", "null"] }
        lastBonusPercentage: { type: ["number", "null"] }
        currentPensionPercentage: { type: ["number", "null"] }
        referenceOneName: { type: ["string", "null"] }
        referenceOneCompany: { type: ["string", "null"] }
        referenceOneJobTitle: { type: ["string", "null"] }
        referenceOneEmail: { type: ["string", "null"] }
        referenceOnePhone: { type: ["string", "null"] }
        referenceTwoName: { type: ["string", "null"] }
        referenceTwoCompany: { type: ["string", "null"] }
        referenceTwoJobTitle: { type: ["string", "null"] }
        referenceTwoEmail: { type: ["string", "null"] }
        referenceTwoPhone: { type: ["string", "null"] }
        rightToWorkStatus:
          type: ["string", "null"]
          enum: [uk_citizen, settled, work_visa, sponsorship_required, unknown, null]
        coverLetter: { type: ["string", "null"] }
        documents:
          type: array
          items: { $ref: "#/components/schemas/CreateCandidateDocumentBody" }
      required: [firstName, lastName, email]
    PublicApplicationResponse:
      type: object
      properties:
        id: { type: integer }
        message: { type: string }
      required: [id, message]
    PublicJobsUnavailableResponse:
      type: object
      properties:
        error: { type: string }
        retryable: { type: boolean }
      required: [error, retryable]
...